requirement.txt
discord
discord.py
discord.py[voice]
python-dotenv
gtts
mutagen
llm-rs
youtube-dl